refactored Shared.Infrastructure persistence
