+++ src/Modules/ClinicManagement/DigiDent.ClinicManagement.EFCorePersistence/Employees/Schedule/WorkingDayConfiguration.cs
@@ -2,7 +2,7 @@
 using DigiDent.ClinicManagement.Domain.Employees.Shared.ValueObjects;
 using DigiDent.ClinicManagement.Domain.Employees.Shared.ValueObjects.Ids;
 using DigiDent.ClinicManagement.EFCorePersistence.Shared.Configurations;
-using DigiDent.Shared.Infrastructure.EfCore.Configurations;
+using DigiDent.Shared.Infrastructure.Persistence.EfCore.Configurations;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace DigiDent.ClinicManagement.EFCorePersistence.Employees.Schedule;
