--- src/Shared/DigiDent.Shared.Abstractions/Factories/IDbConnectionFactory.cs
@@ -1,10 +1,0 @@
-﻿using System.Data;
-
-namespace DigiDent.Shared.Abstractions.Factories;
-
-public interface IDbConnectionFactory<out TConnection>
-    where TConnection: IDbConnection
-{
-    TConnection CreateConnection();
-    TConnection CreateOpenConnection();
-}
