+++ src/Shared/DigiDent.Shared.Infrastructure/Persistence/EfCore/Configurations/AggregateRootConfiguration.cs
@@ -1,7 +1,7 @@
 ﻿using DigiDent.Shared.Kernel.Abstractions;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace DigiDent.Shared.Infrastructure.EfCore.Configurations;
+namespace DigiDent.Shared.Infrastructure.Persistence.EfCore.Configurations;
 
 /// <summary>
 /// Base class for aggregate root configuration.
